Comsolidate helper types
